Add remaining C++ files from debug branch
Molecular core:
- EwaldGrid.h: GPU Ewald grid shift for CPU/LAMMPS agreement (cdf029af)
- GOpt.h, GlobalOptimizer.h: Optimization improvements
- MMFFBuilder.h: Builder updates for groups (4352ce67)
- libMMFF.h: Library interface updates
- MolWorld_sp3_multi.h: Multi-molecule world updates
- MolWorld_sp3_simple.h: Simple world cleanup
- MolecularDatabase.h: Database updates

OpenCL:
- OCL_MM.h: OpenCL molecular mechanics (4eab40b8)
- relax_multi.cl: Multi-molecule relaxation kernel

GUI/SDL:
- MolGUI.h: GUI updates for scan_constr (191694c9)
- MolecularDraw.h: Drawing updates
- MolGUIapp_argv.h: App argument handling

Other:
- IO_utils.h: I/O utilities
- MMFFmulti_lib.cpp: Multi-molecule library

+++ cpp/common/IO_utils.h
@@ -431,8 +431,11 @@ inline int save_npy(const char *fname, int ndims, int* shape, const char *data, 
 class NumpyFile{ public:
     int ndims; 
     int shape[8]; 
-    int ntot;
-    int nBytePerElement;
+    //int ntot;
+    //int nBytePerElement;
+   // char dtype[8];
+    long long ntot;
+    long long nBytePerElement;
     char dtype[8];
     char* data=0;
